Split GL renderer internals
